==> 5_Process HPC outputs.R <==
library(raster); library(mgcv); library(viridisLite); library(rgdal)


setwd("C:/Users/b284718/Google Drive/MSc Project ideas/Nats Pips/Bat Conservation Trust")
k1 <-5
k2 <- 6
load("Natpip_clip2_k1_6_k2_6.RData")

# Check for convergence - ideally Rhat for all nodes <1.05 
hist(model.fit$BUGSoutput$summary[,"Rhat"],breaks=40)
sum(model.fit$BUGSoutput$summary[,"Rhat"]>1.05 )/ nrow(model.fit$BUGSoutput$summary)

# Explore model outputs:
# Posterior distribution of p_d 
p_d # recall true value of detect prob
p_d_samples <- 1/(1+exp(-(model.fit$BUGSoutput$sims.list$eta))) #extract posterior samples of eta, inverse logit transformed to p_d

hist(p_d_samples,breaks=30)

# Explore model-predicted patterns:  

b.samps <- model.fit$BUGSoutput$mean$b #get posterior means of the b matrix (ie the fitted spline basis)

est.counts <- exp(X%*%b.samps) # generate the model estimated true counts - matrix product of spline basis and X matrix generated by jagam, inverse log transformed 

#look at correlation between observed and modelled counts
plot(est.counts[1:n1],site_obsdata$count,xlim=c(0,60),ylim=c(0,60))

# look at posteriors of model parameters 
model.fit.mcmc <- as.mcmc(model.fit)
densplot(model.fit.mcmc)
#convert this to a data frame
model.fit.mat <- as.matrix(model.fit.mcmc)
model.fit.df <- as.data.frame(model.fit.mat)
head(model.fit.df)
# these columns are the regression coefficients
colnames(model.fit.df[, c(281:283, 285:313)])


caterplot(model.fit.df[, c(281:283, 285:313)])

# Now create out object to plot the estimated counts across space and time
maps <- data.frame(x=jagdata$east,y=jagdata$north,est.counts,month=jagdata$month)

colsc <- magma(40)
months <- max(site_obsdata$month)
month_names <- c('April', 'May', 'June', 'July', 'August', 'September', 'October')
# Make a plot for each month
for(k in 1:months){
  map1 <- subset(maps,maps$month==k)
 bks <- seq(0,4,length.out=40) #color scheme for predicted counts, standardised so that the max est count is always 4
 month_name <- month_names[k]
 plot(map1$x,map1$y,pch=16,col=colsc[findInterval(log(map1$est.counts+1),bks)],cex=1.5,main=month_name,axes=F,xlab="",ylab="")
 points(site_obsdata$east, site_obsdata$north, pch = 4, col = 'white')
 par(xpd=T)
  xleg <- seq(-2,0,length.out=10)
  points(x= xleg,y=rep(-1.6,length(xleg)),pch=16,cex=2, col= colsc[floor(seq(1,length(colsc),length.out=length(xleg)))])
  text(x= c(-2,0),y=rep(-1.7,1.7),c("0",round(max(bks),1)),pos=1)
  text(x= c(-1),y=-1.8,c("Log (Est. count)"),pos=1)
}

# get summary of all regression coefficients
rownames(model.fit$BUGSoutput$summary)
model.fit$BUGSoutput$summary[c(217:228),]

# look at the posteriors for beta values
1/(1+exp(-(model.fit$BUGSoutput$mean$beta1)))
1/(1+exp(-(model.fit$BUGSoutput$sd$beta1)))

1/(1+exp(-(model.fit$BUGSoutput$mean$beta2)))
1/(1+exp(-(model.fit$BUGSoutput$sd$beta2)))

1/(1+exp(-(model.fit$BUGSoutput$mean$beta3)))
1/(1+exp(-(model.fit$BUGSoutput$sd$beta3)))

1/(1+exp(-(model.fit$BUGSoutput$mean$lcm_class.fac)))
